reorganized lib and example to follow common folder structure, started unit testing linked list
