+++ lib/pool/linkedlist.mjs
